Add very early exit when required GUCs not set
Previously having incomplete GUC settings would not cause "any" harm
(except performance, where there would be constant attempts to connect
even though the GUCs are invalid, causing connection failure), but in
the desire to be zero-impact when pg_logfebe is unconfigured makes an
early-exit attractive.  So, implement it.

Also add some paranoia around NULL values for some of the GUCs.

Also, set up the GUCs correctly.  How they worked at all to begin with
is a bit a mystery.

Signed-off-by: Daniel Farina <daniel@heroku.com>