Invalidate an open socket if it is inherited from the postmaster
Inheriting this file descriptor would cause a protocol corruption of
the communication with pg_logplexcollector, causing a single dropped
log record.

This manifests as an EPIPE when writing to the socket for the first
time after a fork from postmaster, triggering a re-opening of the
socket, after which symptoms desist.

Since there are several other values that need re-setting because of
inheritance across a fork, hoist that cache-invalidation out of
"fmtLogMsg" to avoid loading "fmtLogMsg" with side effects and teach
that invalidation code to close the socket should it be connected.